feat: Fix None errors in endpoints

+++ euphoria/backend/api/crud/base.py
@@ -60,7 +60,9 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         return db_obj
 
     def delete(self, db: Session, *, id: int) -> ModelType:
-        obj = db.query(self.model).get(id)
-        db.delete(obj)
-        db.commit()
-        return obj
+        if (obj := db.query(self.model).filter(self.model.id == id).first()):
+            db.delete(obj)
+            db.commit()
+            return obj
+        else:
+            return None
